requirements.txt
requests
geoip2
pyyaml